requirements.txt
PyQt6==6.5.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.7.2
flask==2.3.3
python-dotenv==1.0.0 
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.20
httpx==0.25.0